+++ util/mock_env.h
@@ -39,10 +39,6 @@ class MockEnv : public EnvWrapper {
                                  unique_ptr<WritableFile>* result,
                                  const EnvOptions& env_options) override;
 
-  virtual Status NewRandomRWFile(const std::string& fname,
-                                 unique_ptr<RandomRWFile>* result,
-                                 const EnvOptions& options) override;
-
   virtual Status NewDirectory(const std::string& name,
                               unique_ptr<Directory>* result) override;